requirements.txt
# Dependencias principales
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
boto3>=1.26.0
requests>=2.28.0
python-dotenv>=0.19.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0
moto>=2.2.0  # Para simular servicios AWS en tests

# AWS CDK
aws-cdk-lib>=2.0.0
constructs>=10.0.0

# Utilidades
python-jose==3.3.0
passlib==1.7.4
python-multipart>=0.0.5

# Base de datos
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0 